fix(lab3): fix build errors on Windows-Release

+++ src/lab3/display.cpp
@@ -64,7 +64,7 @@ void sdl_display::show_frame(const sim_output& tracks) {
     // Треки
     SDL_SetRenderDrawColor(rend, 255, 255, 255, 255);
     for (size_t p = 0; p < tracks.particle_count; p++) {
-        SDL_RenderDrawLinesF(rend, tracks.tracks[p], tracks.track_len[p]);
+        SDL_RenderDrawLinesF(rend, tracks.tracks[p], (int)tracks.track_len[p]);
     }
 
     // Передаем на экран
